requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.5.0
flask>=2.3.0
werkzeug>=2.3.0

